Refactor RecordSerializer to a public class
